refactor: move schema definitions from main.py to schemas.py and add MonthlyReadingsResponse
