+++ main.py
@@ -4,7 +4,6 @@ from typing import List, Optional
 import models, schemas
 from database import SessionLocal, engine
 from datetime import date
-from pydantic import BaseModel
 
 models.Base.metadata.create_all(bind=engine)
 
@@ -22,55 +21,6 @@ def get_db():
     finally:
         db.close()
 
-# Ajout des nouveaux schémas pour la création
-class MonthCreate(BaseModel):
-    name: str
-    translation: str
-    number: int
-
-class DayCreate(BaseModel):
-    date: date
-    special_event: Optional[str] = None
-    month_id: int
-
-class ReadingCreate(BaseModel):
-    period: str
-    content: str
-    author: str
-    source_book: str
-    page: Optional[str] = None
-    reference: Optional[str] = None
-    day_id: int
-
-class BookCreate(BaseModel):
-    title: str
-    author: str
-    url: Optional[str] = None
-
-# Schémas pour la mise à jour
-class MonthUpdate(BaseModel):
-    name: Optional[str] = None
-    translation: Optional[str] = None
-    number: Optional[int] = None
-
-class DayUpdate(BaseModel):
-    date: Optional[date] = None
-    special_event: Optional[str] = None
-    month_id: Optional[int] = None
-
-class ReadingUpdate(BaseModel):
-    period: Optional[str] = None
-    content: Optional[str] = None
-    author: Optional[str] = None
-    source_book: Optional[str] = None
-    page: Optional[str] = None
-    reference: Optional[str] = None
-
-class BookUpdate(BaseModel):
-    title: Optional[str] = None
-    author: Optional[str] = None
-    url: Optional[str] = None
-
 @app.get("/readings/{date}", response_model=List[schemas.Reading])
 def get_readings_by_date(date_str: str, db: Session = Depends(get_db)):
     try:
@@ -82,7 +32,7 @@ def get_readings_by_date(date_str: str, db: Session = Depends(get_db)):
     except ValueError:
         raise HTTPException(status_code=400, detail="Format de date invalide. Utilisez YYYY-MM-DD")
 
-@app.get("/readings/month/{month_name}", response_model=List[schemas.Reading])
+@app.get("/readings/month/{month_name}", response_model=schemas.MonthlyReadingsResponse)
 def get_readings_by_month(month_name: str, db: Session = Depends(get_db)):
     month = db.query(models.Month).filter(models.Month.name == month_name).first()
     if month is None:
@@ -91,7 +41,12 @@ def get_readings_by_month(month_name: str, db: Session = Depends(get_db)):
     readings = []
     for day in month.days:
         readings.extend(day.readings)
-    return readings
+    
+    return {
+        "month": month_name,
+        "count": len(readings),
+        "readings": readings
+    }
 
 @app.get("/events", response_model=List[dict])
 def get_special_events(db: Session = Depends(get_db)):
@@ -104,7 +59,7 @@ def get_books(db: Session = Depends(get_db)):
 
 # Ajout des nouvelles routes pour la création
 @app.post("/months/", response_model=schemas.Month, status_code=status.HTTP_201_CREATED)
-def create_month(month: MonthCreate, db: Session = Depends(get_db)):
+def create_month(month: schemas.MonthCreate, db: Session = Depends(get_db)):
     db_month = models.Month(**month.model_dump())
     db.add(db_month)
     try:
@@ -116,7 +71,7 @@ def create_month(month: MonthCreate, db: Session = Depends(get_db)):
         raise HTTPException(status_code=400, detail=f"Erreur lors de la création du mois: {str(e)}")
 
 @app.post("/days/", response_model=schemas.Day, status_code=status.HTTP_201_CREATED)
-def create_day(day: DayCreate, db: Session = Depends(get_db)):
+def create_day(day: schemas.DayCreate, db: Session = Depends(get_db)):
     # Vérifier si le mois existe
     month = db.query(models.Month).filter(models.Month.id == day.month_id).first()
     if not month:
@@ -133,7 +88,7 @@ def create_day(day: DayCreate, db: Session = Depends(get_db)):
         raise HTTPException(status_code=400, detail=f"Erreur lors de la création du jour: {str(e)}")
 
 @app.post("/readings/", response_model=schemas.Reading, status_code=status.HTTP_201_CREATED)
-def create_reading(reading: ReadingCreate, db: Session = Depends(get_db)):
+def create_reading(reading: schemas.ReadingCreate, db: Session = Depends(get_db)):
     # Vérifier si le jour existe
     day = db.query(models.Day).filter(models.Day.id == reading.day_id).first()
     if not day:
@@ -154,7 +109,7 @@ def create_reading(reading: ReadingCreate, db: Session = Depends(get_db)):
         raise HTTPException(status_code=400, detail=f"Erreur lors de la création de la lecture: {str(e)}")
 
 @app.post("/books/", response_model=schemas.Book, status_code=status.HTTP_201_CREATED)
-def create_book(book: BookCreate, db: Session = Depends(get_db)):
+def create_book(book: schemas.BookCreate, db: Session = Depends(get_db)):
     # Vérifier si le livre existe déjà
     existing_book = db.query(models.Book).filter(
         models.Book.title == book.title,
@@ -175,7 +130,7 @@ def create_book(book: BookCreate, db: Session = Depends(get_db)):
 
 # Routes de mise à jour
 @app.put("/months/{month_id}", response_model=schemas.Month)
-def update_month(month_id: int, month: MonthUpdate, db: Session = Depends(get_db)):
+def update_month(month_id: int, month: schemas.MonthUpdate, db: Session = Depends(get_db)):
     db_month = db.query(models.Month).filter(models.Month.id == month_id).first()
     if not db_month:
         raise HTTPException(status_code=404, detail="Mois non trouvé")
@@ -192,7 +147,7 @@ def update_month(month_id: int, month: MonthUpdate, db: Session = Depends(get_db
         raise HTTPException(status_code=400, detail=f"Erreur lors de la mise à jour: {str(e)}")
 
 @app.put("/days/{day_id}", response_model=schemas.Day)
-def update_day(day_id: int, day: DayUpdate, db: Session = Depends(get_db)):
+def update_day(day_id: int, day: schemas.DayUpdate, db: Session = Depends(get_db)):
     db_day = db.query(models.Day).filter(models.Day.id == day_id).first()
     if not db_day:
         raise HTTPException(status_code=404, detail="Jour non trouvé")
@@ -215,7 +170,7 @@ def update_day(day_id: int, day: DayUpdate, db: Session = Depends(get_db)):
         raise HTTPException(status_code=400, detail=f"Erreur lors de la mise à jour: {str(e)}")
 
 @app.put("/readings/{reading_id}", response_model=schemas.Reading)
-def update_reading(reading_id: int, reading: ReadingUpdate, db: Session = Depends(get_db)):
+def update_reading(reading_id: int, reading: schemas.ReadingUpdate, db: Session = Depends(get_db)):
     db_reading = db.query(models.Reading).filter(models.Reading.id == reading_id).first()
     if not db_reading:
         raise HTTPException(status_code=404, detail="Lecture non trouvée")
@@ -236,7 +191,7 @@ def update_reading(reading_id: int, reading: ReadingUpdate, db: Session = Depend
         raise HTTPException(status_code=400, detail=f"Erreur lors de la mise à jour: {str(e)}")
 
 @app.put("/books/{book_id}", response_model=schemas.Book)
-def update_book(book_id: int, book: BookUpdate, db: Session = Depends(get_db)):
+def update_book(book_id: int, book: schemas.BookUpdate, db: Session = Depends(get_db)):
     db_book = db.query(models.Book).filter(models.Book.id == book_id).first()
     if not db_book:
         raise HTTPException(status_code=404, detail="Livre non trouvé")
@@ -250,4 +205,4 @@ def update_book(book_id: int, book: BookUpdate, db: Session = Depends(get_db)):
         return db_book
     except Exception as e:
         db.rollback()
-        raise HTTPException(status_code=400, detail=f"Erreur lors de la mise à jour: {str(e)}") 
+        raise HTTPException(status_code=400, detail=f"Erreur lors de la mise à jour: {str(e)}")
